+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.21.1-4.0.24] - 2025-06-26
+### Fixed
+- Possible crash when mining a block with a missing loot table (seelderr) [#818]
+
+## [1.21.1-4.0.23] - 2025-06-08
 ### Added
 - New config options for the starlight charger
   - `work_time` - Sets the time of day energy can be gathered (daytime, nighttime, or anytime)